model/ns3sionna/requirements.txt
sionna
sionna-rt
mitsuba
numpy
protobuf
tensorflow[and-cuda]
pyzmq
gputil

# optional for plotting
matplotlib
